[DOP-15872] Implement Group resource
