+++ src/widgets/users/ui/list/userListElementForCreate.tsx
@@ -1,0 +1,17 @@
+import { Datagrid, TextField } from "react-admin";
+
+const UserListElementForCreate = ({ setUser, setUserId }) => {
+    return (
+        <Datagrid
+            rowClick={(id, resource, element) => {
+                setUser(element.username);
+                setUserId(id);
+            }}
+        >
+            <TextField source={"username"} name={"username"} />
+            <TextField source="description" name={"description"} />
+        </Datagrid>
+    );
+};
+
+export default UserListElementForCreate;
